review update: refactor
