+++ tests/cpp-tests/proj.android/gradle.properties
@@ -31,7 +31,7 @@ PROP_TARGET_SDK_VERSION=28
 # Available architextures (armeabi-v7a | arm64-v8a | x86)
 # To build for multiple architexture, use the `:` between them
 # Example - PROP_APP_ABI=armeabi-v7a:arm64-v8a:x86
-PROP_APP_ABI=armeabi-v7a
+PROP_APP_ABI=arm64-v8a
 
 # android native code build type
 # none, native code will never be compiled.
